deselect brought back with read/write error.
Swap Out is working with thread read and write error that needs to be fixed. Code needs to be organized and cleaned. Initially brought back with a hackaround because JCEF framework does not have a straight replacement for ExecuteJavascriptAndReturnValue.

Used JsQuery to make a request from JS to Java with data as the request, however not handling the request back to Js, instead working with the received data.